# contracts/openapi/console-api.yaml
openapi: 3.0.3
info:
  title: Dynamiq Console API
  description: >
    Control-plane API served by console-api (Phalcon). PRD-001 §9.1.
    Every /admin/v1/* route except /me and /companies requires the X-Company-Id header.
  version: "0.1.0"
servers:
  - url: /admin/v1
security:
  - ssoJwt: []
components:
  securitySchemes:
    ssoJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: SSO JWT issued by the existing Phalcon system (RS256 + JWKS), PRD §7.1/7.6.
  parameters:
    XCompanyId:
      name: X-Company-Id
      in: header
      required: true
      schema:
        type: string
      description: Company context for this request; checked against membership (PRD §7.2).
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    Connection:
      type: object
      properties:
        id: { type: string }
        company_id: { type: string }
        name: { type: string }
        type: { type: string, example: dynamiq.connections.OpenAI }
        api_base: { type: string }
        masked_hint: { type: string }
        max_concurrency: { type: integer }
    ConnectionSecretInput:
      type: object
      required: [secret]
      properties:
        secret: { type: string }
...
paths:
...
  /connections:
    get:
      operationId: listConnections
      parameters: [{ $ref: '#/components/parameters/XCompanyId' }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Connection' } }
    post:
      operationId: createConnection
      parameters: [{ $ref: '#/components/parameters/XCompanyId' }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Connection' }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Connection' }
  /connections/{id}:
    parameters:
      - { $ref: '#/components/parameters/XCompanyId' }
      - { $ref: '#/components/parameters/IdParam' }
    get:
      operationId: getConnection
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Connection' } } } }
    patch:
      operationId: updateConnection
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Connection' }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Connection' } } } }
    delete:
      operationId: deleteConnection
      responses:
        "204": { description: Deleted }
  /connections/{id}/secret:
    parameters:
      - { $ref: '#/components/parameters/XCompanyId' }
      - { $ref: '#/components/parameters/IdParam' }
    put:
      operationId: putConnectionSecret
      summary: Forwarded internally to PUT /internal/v1/connections/{id}/secret; never persisted on the PHP side (PRD §7.4).
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConnectionSecretInput' }
      responses:
        "204": { description: Stored }
  /connections/{id}/test:
    parameters:
      - { $ref: '#/components/parameters/XCompanyId' }
      - { $ref: '#/components/parameters/IdParam' }
    post:
      operationId: testConnection
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  detail: { type: string }
...